Parsing array + scene
