+++ get_material
