+++ src/model/objects/lights/mod.rs
@@ -1,0 +1,5 @@
+pub mod light;
+pub mod ambient_light;
+pub mod point_light;
+pub mod parallel_light;
+pub mod spot_light;
