+++ src/model/objects/lights/parallel_light.rs
@@ -1,0 +1,167 @@
+use std::sync::{Arc, RwLock};
+use crate::{model::{materials::color::Color, maths::{hit::Hit, ray::Ray, vec3::Vec3}, scene::Scene}, ui::{prefabs::vector_ui::get_vector_ui, ui::UI, uielement::{Category, UIElement}, utils::misc::{ElemType, Property, Value}}, BOUNCE_OFFSET};
+use super::light::{AnyLight, Light};
+
+#[derive(Debug, Clone)]
+pub struct ParallelLight {
+    dir: Vec3,
+    intensity: f64,
+    color: Color,
+}
+
+impl ParallelLight {
+    // Accessors
+    pub fn dir(&self) -> &Vec3 {
+        &self.dir
+    }
+    pub fn intensity(&self) -> f64 {
+        self.intensity
+    }
+    pub fn color(&self) -> &Color {
+        &self.color
+    }
+
+    // Constructor
+    pub fn new(dir: Vec3, intensity: f64, color: Color) -> Self {
+        self::ParallelLight {
+            dir,
+            intensity,
+            color,
+        }
+    }
+}
+
+impl Light for ParallelLight {
+    fn get_diffuse(&self, hit: &Hit) -> Color {
+        let mut ratio = (-self.dir()).dot(hit.norm());
+        if ratio < 0. {
+            return Color::new(0., 0., 0.);
+        }
+        ratio *= 0_f64.max(self.intensity());
+        (ratio * self.color()).clamp(0., 1.)
+    }
+
+    fn get_specular(&self, hit: &Hit, ray: &Ray) -> Color {
+        let to_light = -self.dir();
+        let reflected = (-(&to_light) - hit.norm().dot(&-to_light) * 2. * hit.norm()).normalize();
+        let mut ratio = (-ray.get_dir()).normalize().dot(&reflected);
+        if ratio < 0. {
+            return Color::new(0., 0., 0.);
+        }
+        ratio = ratio.powf(50.);
+        ratio *= self.intensity().powi(2);
+        (ratio * self.color()).clamp(0., 1.)
+    }
+
+    fn is_shadowed(&self, scene: &Scene, hit: &Hit) -> bool {
+        let shadow_ray = Ray::new(hit.pos() + hit.norm() * BOUNCE_OFFSET, -self.dir(), 0);
+        for element in scene.elements() {
+            if let Some(t) = element.shape().intersect(&shadow_ray) {
+                for t in t {
+                    if t > 0. {
+                        return true;
+                    }
+                }
+            }
+        }
+        false
+    }
+
+    fn as_parallel_light(&self) -> Option<&ParallelLight> {
+        Some(self)
+    }
+
+    fn as_parallel_light_mut(&mut self) -> Option<&mut ParallelLight> {
+        Some(self)
+    }
+
+	fn get_ui(&self, light: &AnyLight, ui: &mut UI, scene: &Arc<RwLock<Scene>>) -> UIElement {
+		let id = light.id().clone();
+		let mut category = UIElement::new(format!("Parallel light {}", id).as_str(), format!("light{}", id).as_str(), ElemType::Category(Category::collapsed()), ui.uisettings());
+
+		let dir = get_vector_ui(*light.light().as_parallel_light().unwrap().dir(), "Direction", "dir", ui.uisettings(), 
+		Box::new(move |_, value, scene, _| {
+			let mut scene = scene.write().unwrap();
+			let light = scene.light_mut_by_id(id.clone()).unwrap();
+			if let Some(light) = light.light_mut().as_parallel_light_mut() {
+				if let Value::Float(value) = value {
+					light.dir.set_x(value);
+					light.dir = light.dir.normalize();
+				}
+			}
+		}),
+		Box::new(move |_, value, scene, _| {
+			let mut scene = scene.write().unwrap();
+			let light = scene.light_mut_by_id(id.clone()).unwrap();
+			if let Some(light) = light.light_mut().as_parallel_light_mut() {
+				if let Value::Float(value) = value {
+					light.dir.set_y(value);
+					light.dir = light.dir.normalize();
+				}
+			}
+		}),
+		Box::new(move |_, value, scene, _| {
+			let mut scene = scene.write().unwrap();
+			let light = scene.light_mut_by_id(id.clone()).unwrap();
+			if let Some(light) = light.light_mut().as_parallel_light_mut() {
+				if let Value::Float(value) = value {
+					light.dir.set_z(value);
+					light.dir = light.dir.normalize();
+				}
+			}
+		}), false, None, None);
+
+		let color = get_vector_ui(light.light().as_parallel_light().unwrap().color().to_vec3(), "Color", "color", ui.uisettings(), 
+		Box::new(move |_, value, scene, _| {
+			let mut scene = scene.write().unwrap();
+			let light = scene.light_mut_by_id(id.clone()).unwrap();
+			if let Some(light) = light.light_mut().as_parallel_light_mut() {
+				if let Value::Float(value) = value {
+					light.color = Color::new(value, light.color.g(), light.color.b());
+				}
+			}
+		}),
+		Box::new(move |_, value, scene, _| {
+			let mut scene = scene.write().unwrap();
+			let light = scene.light_mut_by_id(id.clone()).unwrap();
+			if let Some(light) = light.light_mut().as_parallel_light_mut() {
+				if let Value::Float(value) = value {
+					light.color = Color::new(light.color.r(), value, light.color.b());
+				}
+			}
+		}),
+		Box::new(move |_, value, scene, _| {
+			let mut scene = scene.write().unwrap();
+			let light = scene.light_mut_by_id(id.clone()).unwrap();
+			if let Some(light) = light.light_mut().as_parallel_light_mut() {
+				if let Value::Float(value) = value {
+					light.color = Color::new(light.color.r(), light.color.g(), value);
+				}
+			}
+		}), true, Some(0.), Some(1.));
+
+		let intensity = UIElement::new("Intensity", "intensity", ElemType::Property(Property::new(Value::Float(light.light().as_parallel_light().unwrap().intensity()),
+		Box::new(move |_, value, scene, _| {
+			let mut scene = scene.write().unwrap();
+			let light = scene.light_mut_by_id(id.clone()).unwrap();
+			if let Some(light) = light.light_mut().as_parallel_light_mut() {
+				if let Value::Float(value) = value {
+					light.intensity = value;
+				}
+			}
+		}),
+		Box::new(move |value, _, _| {
+			if let Value::Float(value) = value {
+				if *value < 0. {
+					return Err("The value should not be inferior to 0".to_string());
+				}
+			}
+			Ok(())
+		}), ui.uisettings())), ui.uisettings());
+
+		category.add_element(dir);
+		category.add_element(color);
+		category.add_element(intensity);
+		category
+	}
+}
